% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{package}
\name{convlog-package}
\alias{convlog}
\alias{convlog-package}
\title{convlog: Read Mahjong Logs From 'tenhou.net/6' Format}
\description{
Offers wrappers for the 'convlog' crate from 'mjai-reviewer' (\url{https://github.com/Equim-chan/mjai-reviewer}) that can directly read mahjong logs from 'tenhou.net/6' format into tibbles.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Equim-chan (Author of the 'convlog' crate) [copyright holder]
}

}
\keyword{interal}
