% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read-tenhou6}
\alias{read-tenhou6}
\alias{read_tenhou6}
\alias{read_remote_mjlog}
\alias{read_mjlog}
\title{Read and parse 'tenhou.net/6' format log}
\usage{
read_tenhou6(file, .progress = FALSE)

read_remote_mjlog(logid, .progress = FALSE)

read_mjlog(file, .progress = FALSE)
}
\arguments{
\item{file}{A character vector.
This argument is simply passed to \code{scan()},
so each element can be either a path to a local file or a URL.}

\item{.progress}{Whether to show progress bar for \code{purrr::map_chr()}.}

\item{logid}{A character vector that represents identifiers of log files.}
}
\value{
A named list that contains following elements:
\itemize{
\item \code{game_info}: A tibble that contains information about the games.
\item \code{round_info}: A tibble that contains information about rounds.
\item \code{paifu}: A tibble that represents paifu.
}
}
\description{
Read and parse 'tenhou.net/6' format JSON files
while transforming them into 'mjai' format.
}
\details{
\code{read_remote_mjlog()} internally reads
remote JSON files corresponding to \code{logid},
and converts them into the same format as \code{read_tenhou6()}.
Note that \code{read_remote_mjlog()} is rate-limited to 2 requests per second
to access the server.

Alternatively, \code{read_mjlog()} can directly read local 'MJLOG' XML files
while converting them into 'mjai' format.
This function returns almost the same result as \code{read_tenhou6()},
but they are not exactly the same.
As far as I have noticed, the differences are:
\itemize{
\item \code{tehais} in \code{game_info} are not arranged.
\item \code{reach_accepted} events are always inserted
immediately after the reach event in this implementation,
even when the reach indicator was melded by another player.
Due to this, numbering style of \code{round_id} is also different
than in \code{read_tenhou6()}.
\item \code{tsumogiri} detection is much stricter than in \code{read_tenhou6()}.
Even after melds, if the discard is the same as the previous draw,
it is considered a tsumogiri.
\item \code{ura_markers} are not revealed when there is no "doraHaiUra" attribute.
}
}
